requirements.txt
# pip install -r requirements.txt
pytest==4.2.0
Flask==1.0.2
newsapi-python==0.2.3